[fix]: Datatype return

+++ 9-binary_tree_height.c
@@ -7,7 +7,7 @@
 
 size_t binary_tree_height(const binary_tree_t *tree)
 {
-    int right = 0, left = 0;
+    size_t right = 0, left = 0;
 
     if (!tree || (!tree->right && !tree->left))
 	return (0);
